Pod control software for Raspberry Pi
This is what the PI is going to be running while on the pod. It has a bunch of variables for the different sensors and inputs. It also has a bunch of constants that regulate the logic in the state changes (makes it easy to tweak). Some of the functions are empty right now like readGUI which is supposed to read in the data from the Master-Arduino and update the different variables. I'll update this tomorrow, but I figure I'll put up what I have tonight.